% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_Exact_Test.R
\name{Fisher_Exact_Test}
\alias{Fisher_Exact_Test}
\title{Estimates Fisher's Exact P-values to test Additive Sharp Null Hypotheses}
\usage{
Fisher_Exact_Test(fun, W, Y_obs, N, H0)
}
\arguments{
\item{fun}{Specifies the function that will be used for the test statistic. Options are "Diff", "Log", "Median", "T-statistic", and "Kolmogorov-Smirnov". See user manual for further details.}

\item{W}{A vector that indicates treatment status. W takes the value of 1 for units that were treated and zero otherwise.}

\item{Y_obs}{A vector with the observed outcomes of the variable of interest.}

\item{N}{Number of randomly simulated treatment vectors used to obtain the distribution of the test statistic.}

\item{H0}{Additive sharp null hypothesis that will be tested. If H0 = c, we will test the null that the treatment effect is equal to c for all units.}
}
\value{
A list containing two elements. The first element specifies the p-value that we obtain from the hypothesis test. The second element of the list provides an approximation of the distribution of the test statistic.
}
\description{
Estimates Fisher's Exact P-values to test Additive Sharp Null Hypotheses
}
\examples{
pooled_hh <- read_dta("/data_modified/data_modified/pooled_hh.dta")

pooled_hh <- pooled_hh[!(pooled_hh$ctotal_pcmonth_end == 0),]
pooled_hh <- pooled_hh[!(pooled_hh$cfood_pcmonth_end== 0),]

pooled_hh$Expenditure <- log(pooled_hh$ctotal_pcmonth_end)

pooled_hh$Food_Expenditure <- log(pooled_hh$cfood_pcmonth_end)

pooled_hh <- pooled_hh[!(is.na(pooled_hh$treatment)),]

pooled_hh <- pooled_hh[!(is.na(pooled_hh$Expenditure)),]

pooled_hh <- pooled_hh[!(is.na(pooled_hh$Food_Expenditure)),]

Y <- pooled_hh$Expenditure
W <- pooled_hh$treatment

###########Test H0: treatment effect = 0 for all units with T(W,Y_obs) = T^{diff}
First_Test <- Fisher_Exact_Test(fun = "Diff", W=W, Y_obs = Y, N = 1000, H0 = 0)


##########Test H0: treatment effect on household's food expenditure was equal to 1 for all observations
Y_2 <- pooled_hh$Food_Expenditure

W <- pooled_hh$treatment

Second_Test <- Fisher_Exact_Test(fun = "Kolmogorov-Smirnov", W=W, Y_obs = Y_2, N = 1000, H0 = 1)

print("the p-value is equal to")

Second_Test[1]




}
