% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_Plot.R
\name{Fisher_Plot}
\alias{Fisher_Plot}
\title{Plots histograms and kernel densities of the distribution of the test statistic.}
\usage{
Fisher_Plot(output, plot_type)
}
\arguments{
\item{output}{An object of class "Fisher_Exact_Test" or "Multiple_Tests". Takes in the output of the function Fisher_Exact_Test() or Multiple_Tests().}

\item{plot_type}{Specifies if the plot will be a histogram or a kernel density. Options are "Histogram" and "Kernel".}
}
\value{
Returns a plot of the specified type. The purple line marks the value of the observed test statistic.
}
\description{
Plots histograms and kernel densities of the distribution of the test statistic.
}
\examples{
###This example is a continuation of the example for the documentation of the function "Fisher_Exact_Test"

Fisher_Plot(First_Test,plot_type = "Histogram")

Fisher_Plot(First_Test,plot_type = "Kernel")
}
