% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_Test_Reg_Coeff.R
\name{Fisher_Test_Reg_Coeff}
\alias{Fisher_Test_Reg_Coeff}
\title{Estimates Estimates Fisher's Exact P-values to test Additive Sharp Null Hypotheses with Covariates Using the Statistic T_reg_coeff}
\usage{
Fisher_Test_Reg_Coeff(W, Y_obs, N, H0, x)
}
\arguments{
\item{W}{A vector that indicates treatment status. W takes the value of 1 for units that were treated and zero otherwise.}

\item{Y_obs}{A vector with the observed outcomes of the variable of interest.}

\item{N}{Number of randomly simulated treatment vectors used to obtain the distribution of the test statistic.}

\item{H0}{Additive sharp null hypothesis that will be tested. If H0 = c, we will test the null that the treatment effect is equal to c for all units.}

\item{x}{A matrix of covariates to be used as control variables.}
}
\value{
A list containing two elements. The first element specifies the p-value that we obtain from the hypothesis test. The second element of the list provides an approximation of the distribution of the test statistic.
}
\description{
Estimates Estimates Fisher's Exact P-values to test Additive Sharp Null Hypotheses with Covariates Using the Statistic T_reg_coeff. See user's manual for a more detailed description.
}
\examples{

#This example is a continuation of the one provided in the documentation for the 'Fisher_Exact_Test' function

Z <- as.matrix(cbind(pooled_hh$asset_tot_value_bsl,pooled_hh$ctotal_pcmonth_bsl,pooled_hh$cfood_pcmonth_bsl))

Fisher_Test_Reg_Coeff(W=W, Y_obs = Y, N = 500, H0 = 0, x = Z)

}
