% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple_Tests.R
\name{Multiple_Tests}
\alias{Multiple_Tests}
\title{Combines two different test statistics to perform a joint hypothesis test.}
\usage{
Multiple_Tests(test_1, test_2, fun)
}
\arguments{
\item{test_1}{An object of class "Fisher_Exact_Test". Takes in the output of the function Fisher_Exact_Test.}

\item{test_2}{An object of class "Fisher_Exact_Test". Takes in the output of the function Fisher_Exact_Test.}

\item{fun}{The function for the test statistic. Options are "Mean" and "Max". See user manual for more information.}
}
\value{
A list containing two elements. The first element specifies the p-value that we obtain from the hypothesis test. The second element of the list provides an approximation of the distribution of the test statistic.
}
\description{
Combines two different test statistics to perform a joint hypothesis test.
}
\examples{
###This example is a continuation of the example for the documentation of the function "Fisher_Exact_Test"

First_Test <- Fisher_Exact_Test(fun = "Diff", W=W, Y_obs = Y, N = 1000, H0 = 0)

Second_Test <- Fisher_Exact_Test(fun = "Kolmogorov-Smirnov", W=W, Y_obs = Y_2, N = 1000, H0 = 1)

Multiple_Tests(First_Test,Second_Test, fun = "Mean")

}
